docker/deepseek-ocr-service/requirements.txt
# DeepSeek-OCR Service Dependencies

# Core ML
torch>=2.4.0
transformers>=4.45.0
accelerate>=0.34.0

# API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
python-multipart>=0.0.12

# Image processing
Pillow>=10.4.0

# Utilities
requests>=2.32.0

# flash_attn устанавливается отдельно в Dockerfile
# (может не собраться для некоторых GPU архитектур)
